Remove deprecated scripts
