+++ routes/home_routes.py
@@ -1,4 +1,5 @@
 """Homepage routes for DreamArtMachine."""
+
 from __future__ import annotations
 
 from flask import Blueprint, redirect, render_template, url_for
@@ -19,3 +20,24 @@ def root() -> "Response":
 def home() -> str:
     """Render the application homepage."""
     return render_template("home.html")
+
+
+@bp.route("/upload")
+@login_required
+def upload() -> str:
+    """Render the artwork upload page."""
+    return render_template("upload.html")
+
+
+@bp.route("/artworks")
+@login_required
+def artworks() -> str:
+    """Render the artworks listing page."""
+    return render_template("artworks.html")
+
+
+@bp.route("/finalised")
+@login_required
+def finalised() -> str:
+    """Render the finalised artworks page."""
+    return render_template("finalised.html")
