+++ templates/admin/mockups.html
@@ -2,123 +2,8 @@
 {% block title %}Mockup Admin{% endblock %}
 {% block content %}
 <h1>Mockup Admin Dashboard</h1>
-<p class="page-description">Upload new blank mockups, create categories, and use AI to automatically sort them.</p>
+<p class="page-description">Upload and organise mockup files.</p>
 <div class="main-content">
-
-    <div class="controls-bar">
-        <div class="form-group">
-            <label for="aspect-ratio-selector">Aspect Ratio:</label>
-            <select id="aspect-ratio-selector">
-                {% for ar in aspect_ratios %}
-                <option value="{{ ar }}" {% if ar == current_aspect %}selected{% endif %}>{{ ar }}</option>
-                {% endfor %}
-            </select>
-        </div>
-        <div class="form-group">
-            <label for="per-page-selector">Show:</label>
-            <select id="per-page-selector">
-                <option value="10" {% if per_page == 10 %}selected{% endif %}>10</option>
-                <option value="20" {% if per_page == 20 %}selected{% endif %}>20</option>
-                <option value="50" {% if per_page == 50 %}selected{% endif %}>50</option>
-            </select>
-        </div>
-        <div class="form-group">
-            <label for="sort-selector">Sort by:</label>
-            <select id="sort-selector">
-                <option value="name" {% if sort_by == 'name' %}selected{% endif %}>Name</option>
-                <option value="date" {% if sort_by == 'date' %}selected{% endif %}>Date Added</option>
-            </select>
-        </div>
-        <button id="find-duplicates-btn" class="btn btn-secondary" data-url="{{ url_for('mockup_admin.find_duplicates', aspect=current_aspect) }}">Find Duplicates</button>
-    </div>
-
-    <div class="admin-grid">
-        <section class="admin-section">
-            <h2>Manage Categories for {{ current_aspect }}</h2>
-            <form action="{{ url_for('mockup_admin.create_category', aspect=current_aspect) }}" method="POST" class="form-inline">
-                <input type="text" name="category_name" placeholder="New category name" required class="form-control-large">
-                <button type="submit" class="btn btn-secondary btn-sm">Create</button>
-            </form>
-            <div class="category-list">
-                <a href="{{ url_for('mockup_admin.dashboard', aspect=current_aspect) }}" class="category-tag {% if category_filter == 'All' %}active{% endif %}">All Mockups</a>
-                <a href="{{ url_for('mockup_admin.dashboard', aspect=current_aspect, category='Uncategorised') }}" class="category-tag {% if category_filter == 'Uncategorised' %}active{% endif %}">Uncategorised</a>
-                {% for category in categories %}
-                <a href="{{ url_for('mockup_admin.dashboard', aspect=current_aspect, category=category) }}" class="category-tag {% if category_filter == category %}active{% endif %}">{{ category }}</a>
-                {% endfor %}
-            </div>
-        </section>
-
-        <section class="admin-section">
-            <h2>Upload New Mockups to {{ current_aspect }}</h2>
-             <form id="upload-form" method="post" action="{{ url_for('mockup_admin.upload_mockup', aspect=current_aspect) }}" enctype="multipart/form-data">
-                <input id="file-input" type="file" name="mockup_files" accept="image/*" multiple hidden>
-                <div id="dropzone" class="upload-dropzone">
-                    Drag & Drop new mockups here or click to choose files
-                </div>
-            </form>
-        </section>
-    </div>
-
-    <section class="admin-section">
-        <h2>Showing {{ category_filter }} Mockups ({{ mockups|length }} of {{ total_mockups }})</h2>
-        <div class="artwork-grid" id="mockup-grid">
-            {% for mockup in mockups %}
-            <div class="gallery-card" data-filename="{{ mockup.filename }}" data-category="{{ mockup.category }}">
-                <div class="card-thumb">
-                    <img class="card-img-top" 
-                         src="{{ url_for('mockup_admin.mockup_thumbnail', aspect=current_aspect, filename=mockup.filename) }}" 
-                         data-fullsize-url="{{ url_for('mockup_admin.mockup_image', aspect=current_aspect, category=mockup.category, filename=mockup.filename) }}"
-                         alt="{{ mockup.filename }}">
-                </div>
-                <div class="card-details">
-                    <p class="card-title" style="font-weight: bold;">{{ mockup.filename }}</p>
-                    <div class="category-badge {% if mockup.category == 'Uncategorised' %}uncategorised{% endif %}">{{ mockup.category }}</div>
-                    <div class="categorize-actions">
-                        <select class="form-control">
-                            <option value="">-- Change Category --</option>
-                            {% for category in categories %}
-                            <option value="{{ category }}" {% if category == mockup.category %}selected{% endif %}>{{ category }}</option>
-                            {% endfor %}
-                        </select>
-                        <div class="button-row" style="margin-top: 0.5rem;">
-                            <button class="btn btn-secondary btn-categorize">Suggest w/ AI</button>
-                            <button class="btn btn-primary btn-save-move">Save</button>
-                        </div>
-                    </div>
-                    <button class="btn btn-danger btn-delete" style="margin-top: 0.5rem; width: 100%;">Delete</button>
-                </div>
-                <div class="card-overlay hidden"></div>
-            </div>
-            {% endfor %}
-        </div>
-        {% if not mockups %}
-        <p>No mockups found for this filter.</p>
-        {% endif %}
-    </section>
-
-    <nav class="pagination">
-        {% if page > 1 %}
-        <a href="{{ url_for('mockup_admin.dashboard', aspect=current_aspect, page=page-1, per_page=per_page, category=category_filter, sort=sort_by) }}" class="btn btn-secondary">&laquo; Previous</a>
-        {% endif %}
-        <span>Page {{ page }} of {{ total_pages }}</span>
-        {% if page < total_pages %}
-        <a href="{{ url_for('mockup_admin.dashboard', aspect=current_aspect, page=page+1, per_page=per_page, category=category_filter, sort=sort_by) }}" class="btn btn-secondary">Next &raquo;</a>
-        {% endif %}
-    </nav>
-  </div>
-
-  <div id="upload-modal" class="analysis-modal"></div>
-<div id="image-modal" class="modal-bg" style="display: none;"><span class="modal-close">&times;</span><img src="" alt="Full size mockup" class="modal-img"></div>
-<div id="duplicates-modal" class="analysis-modal" style="display: none;">
-    <div class="analysis-box">
-        <button class="modal-close">&times;</button>
-        <h3>Duplicate Images Found</h3>
-        <p>The following images appear to be visually identical.</p>
-        <div id="duplicates-list" style="text-align: left; font-size: 0.9em; max-height: 60vh; overflow-y: auto;"></div>
-    </div>
+  <p>Mockup management tools are coming soon.</p>
 </div>
-
-{% endblock %}
-{% block scripts %}
-<script src="{{ url_for('static', filename='js/mockup-admin.js') }}"></script>
 {% endblock %}
