+++ templates/artworks.html
@@ -1,110 +1,9 @@
-{# Use blueprint-prefixed endpoints like 'artwork.home' in url_for #}
 {% extends "main.html" %}
 {% block title %}Artwork | ArtNarrator{% endblock %}
 {% block content %}
 <h1><img src="{{ url_for('static', filename='icons/svg/light/number-circle-two-light.svg') }}" class="hero-step-icon" alt="Step 2: Artwork" />Artwork</h1>
-<p class="page-description">Manage artwork through the analysis pipeline.</p>
+<p class="page-description">Browse all artworks currently uploaded.</p>
 <div class="main-content">
-<div class="gallery-section">
-
-  {% if ready_artworks %}
-    <h2 class="mb-3">Ready to Analyze</h2>
-    <div class="artwork-grid">
-      {% for art in ready_artworks %}
-      <div class="gallery-card" data-filename="{{ art.filename }}" data-aspect="{{ art.aspect }}" data-base="{{ art.base }}">
-        <div class="card-thumb">
-          <img class="card-img-top"
-               src="{{ url_for('artwork.unanalysed_image', filename=art.thumb) }}"
-               alt="{{ art.title }}">
-        </div>
-        <span class="status-icon"></span>
-        <div class="card-details">
-          <div class="card-title">{{ art.title }}</div>
-          <div class="desc-snippet"></div>
-          <div class="button-row">
-            <button class="btn btn-primary btn-analyze" data-provider="openai">Analyze with OpenAI</button>
-            <button class="btn btn-secondary btn-sign" data-base="{{ art.base }}">Sign Artwork</button>
-            <form method="post" action="{{ url_for('artwork.delete_artwork', seo_folder=art.base) }}" style="display:inline;" onsubmit="return confirm('Delete this artwork and all files? This cannot be undone.');">
-              <button type="submit" class="btn btn-danger">Delete</button>
-            </form>
-          </div>
-        </div>
-        <div class="card-overlay hidden"></div>
-      </div>
-      {% endfor %}
-    </div>
-  {% endif %}
-
-  {% if processed_artworks %}
-    <h2 class="mb-3 mt-5">Processed Artworks</h2>
-    <div class="artwork-grid">
-      {% for art in processed_artworks %}
-      <div class="gallery-card" data-filename="{{ art.filename }}" data-aspect="{{ art.aspect }}">
-        <div class="card-thumb">
-          <img class="card-img-top"
-               src="{{ url_for('artwork.processed_image', filename=art.seo_folder ~ '/' ~ art.thumb) }}"
-               alt="{{ art.title }}">
-        </div>
-        <span class="status-icon"></span>
-        <div class="card-details">
-          <div class="card-title">{{ art.title }}</div>
-          <div class="desc-snippet"></div>
-          <div class="button-row">
-            {# --- CORRECTED LINE --- #}
-            <a href="{{ url_for('artwork.edit_listing', aspect=art.aspect, filename=art.seo_folder ~ '.jpg') }}" class="btn btn-primary btn-edit">Review</a>
-            <form method="post" action="{{ url_for('artwork.finalise_artwork', seo_folder=art.seo_folder) }}" style="display:inline;">
-              <button type="submit" class="btn btn-success">Finalise</button>
-            </form>
-            <button class="btn btn-secondary btn-analyze" data-provider="openai">Re-Analyze</button>
-            <form method="post" action="{{ url_for('artwork.delete_artwork', seo_folder=art.seo_folder) }}" style="display:inline;" onsubmit="return confirm('Delete this artwork and all files? This cannot be undone.');">
-              <button type="submit" class="btn btn-danger">Delete</button>
-            </form>
-          </div>
-        </div>
-        <div class="card-overlay hidden"></div>
-      </div>
-      {% endfor %}
-    </div>
-  {% endif %}
-
-  {% if finalised_artworks %}
-    <h2 class="mb-3 mt-5">Finalised Artworks</h2>
-    <div class="artwork-grid">
-      {% for art in finalised_artworks %}
-      <div class="gallery-card" data-filename="{{ art.filename }}" data-aspect="{{ art.aspect }}">
-        <div class="card-thumb">
-          <img class="card-img-top"
-               src="{{ url_for('artwork.finalised_image', filename=art.seo_folder ~ '/' ~ art.thumb) }}"
-               alt="{{ art.title }}">
-        </div>
-        <span class="status-icon"></span>
-        <div class="card-details">
-          <div class="card-title">{{ art.title }}</div>
-          <div class="desc-snippet"></div>
-          <div class="button-row">
-            {# --- CORRECTED LINE --- #}
-            <a href="{{ url_for('artwork.edit_listing', aspect=art.aspect, filename=art.seo_folder ~ '.jpg') }}" class="btn btn-secondary btn-edit">Edit</a>
-            <form method="post" action="{{ url_for('artwork.lock_it_in', seo_folder=art.seo_folder) }}" style="display:inline;">
-              <button type="submit" class="btn btn-primary">Lock It In</button>
-            </form>
-            <button class="btn btn-secondary btn-analyze" data-provider="openai">Re-Analyze</button>
-            <form method="post" action="{{ url_for('artwork.delete_artwork', seo_folder=art.seo_folder) }}" style="display:inline;" onsubmit="return confirm('Delete this artwork and all files? This cannot be undone.');">
-              <button type="submit" class="btn btn-danger">Delete</button>
-            </form>
-          </div>
-        </div>
-        <div class="card-overlay hidden"></div>
-      </div>
-      {% endfor %}
-    </div>
-  {% endif %}
-  {% if not ready_artworks and not processed_artworks and not finalised_artworks %}
-    <p class="empty-msg">No artworks found. Please upload artwork to get started!</p>
-  {% endif %}
-
+  <p>Artwork management features are coming soon.</p>
 </div>
-</div>
-{% endblock %}
-{% block scripts %}
-<script src="{{ url_for('static', filename='js/artworks.js') }}"></script>
 {% endblock %}
