+++ templates/finalised.html
@@ -1,0 +1,9 @@
+{% extends "main.html" %}
+{% block title %}Finalised Artworks{% endblock %}
+{% block content %}
+<h1><img src="{{ url_for('static', filename='icons/svg/light/number-circle-four-light.svg') }}" class="hero-step-icon" alt="Step 4: Finalised" />Finalised Artworks</h1>
+<p class="page-description">Artworks ready for export.</p>
+<div class="main-content">
+  <p>No finalised artworks yet.</p>
+</div>
+{% endblock %}
