+++ templates/index.html
@@ -1,4 +1,4 @@
-{# Use blueprint-prefixed endpoints like 'artwork.home' in url_for #}
+{# Use blueprint-prefixed endpoints like 'home.home' in url_for #}
 {% extends "main.html" %}
 {% block title %}ArtNarrator Home{% endblock %}
 {% block content %}
@@ -6,7 +6,7 @@
 
 <!-- ========== [IN.1] Home Hero Section ========== -->
 <div class="home-hero">
-  <h1><img src="{{ url_for('static', filename='logos/logo-vector.svg') }}" alt="" class="artnarrator-logo"/>Welcome to ArtNarrator Listing Machine</h1>
+  <h1><img src="{{ url_for('static', filename='logos/logo-vector.svg') }}" alt="ArtNarrator Logo" class="artnarrator-logo"/>Welcome to ArtNarrator Listing Machine</h1>
   <p class="home-intro">
     G’day! Start in the Artwork section, analyze new pieces, review your AI-generated listings and mockups, and prep everything for marketplace export—all streamlined for you.
   </p>
@@ -14,16 +14,16 @@
 
 <!-- ========== [IN.2] Home Quick Actions ========== -->
 <div class="workflow-row">
-  <a href="{{ url_for('artwork.upload_artwork') }}" class="workflow-btn">
+  <a href="{{ url_for('home.upload') }}" class="workflow-btn">
     <img src="{{ url_for('static', filename='icons/svg/light/number-circle-one-light.svg') }}" class="step-btn-icon" alt="Step 1" />
     Upload Artwork
   </a>
-  <a href="{{ url_for('artwork.artworks') }}" class="workflow-btn">
+  <a href="{{ url_for('home.artworks') }}" class="workflow-btn">
     <img src="{{ url_for('static', filename='icons/svg/light/number-circle-two-light.svg') }}" class="step-btn-icon" alt="Step 2" />
     Artwork<br>Select to Analyze
   </a>
   {% if latest_artwork %}
-    <a href="{{ url_for('artwork.edit_listing', aspect=latest_artwork.aspect, filename=latest_artwork.filename) }}" class="workflow-btn">
+    <a href="{{ url_for('finalise.edit_listing', slug=latest_artwork.slug) }}" class="workflow-btn">
       <img src="{{ url_for('static', filename='icons/svg/light/number-circle-three-light.svg') }}" class="step-btn-icon" alt="Step 3" />
       Edit Review<br>and Finalise Artwork
     </a>
@@ -33,7 +33,7 @@
       Edit Review<br>and Finalise Artwork
     </span>
   {% endif %}
-  <a href="{{ url_for('artwork.finalised_gallery') }}" class="workflow-btn">
+  <a href="{{ url_for('home.finalised') }}" class="workflow-btn">
     <img src="{{ url_for('static', filename='icons/svg/light/number-circle-four-light.svg') }}" class="step-btn-icon" alt="Step 4" />
     Finalised Gallery<br>(Select to List)
   </a>
