+++ templates/main.html
@@ -53,10 +53,10 @@
             <div class="nav-column">
                 <h3>Artwork &amp; Gallery</h3>
                 <ul>
-                    <li><a href="{{ '#' }}">Upload Artwork</a></li>
-                    <li><a href="{{ '#' }}">All Artworks</a></li>
-                    <li><a href="{{ '#' }}">Finalised</a></li>
-                    <li><a href="{{ '#' }}">Locked</a></li>
+                    <li><a href="{{ url_for('home.upload') }}">Upload Artwork</a></li>
+                    <li><a href="{{ url_for('home.artworks') }}">All Artworks</a></li>
+                    <li><a href="{{ url_for('home.finalised') }}">Finalised</a></li>
+                    <li><a href="#">Locked</a></li>
                 </ul>
             </div>
             <div class="nav-column">
@@ -71,10 +71,10 @@
             <div class="nav-column">
                 <h3>Exports &amp; Admin</h3>
                 <ul>
-                    <li><a href="{{ '#' }}">Sellbrite Management</a></li>
-                    <li><a href="{{ '#' }}">Admin Dashboard</a></li>
-                    <li><a href="{{ '#' }}">Admin Security</a></li>
-                    <li><a href="{{ '#' }}">Description Editor (GDWS)</a></li>
+                    <li><a href="#">Sellbrite Management</a></li>
+                    <li><a href="{{ url_for('admin.dashboard') }}">Admin Dashboard</a></li>
+                    <li><a href="{{ url_for('admin.security_page') }}">Admin Security</a></li>
+                    <li><a href="#">Description Editor (GDWS)</a></li>
                     <li><a href="{{ url_for('login') }}">Login</a></li>
                 </ul>
             </div>
@@ -135,10 +135,10 @@
             <div class="footer-column">
                 <h4>Artwork &amp; Gallery</h4>
                 <ul>
-                    <li><a href="{{ '#' }}">Upload Artwork</a></li>
-                    <li><a href="{{ '#' }}">Artworks</a></li>
-                    <li><a href="{{ '#' }}">Finalised</a></li>
-                    <li><a href="{{ '#' }}">Locked</a></li>
+                    <li><a href="{{ url_for('home.upload') }}">Upload Artwork</a></li>
+                    <li><a href="{{ url_for('home.artworks') }}">Artworks</a></li>
+                    <li><a href="{{ url_for('home.finalised') }}">Finalised</a></li>
+                    <li><a href="#">Locked</a></li>
                 </ul>
             </div>
             <div class="footer-column">
@@ -151,10 +151,10 @@
             <div class="footer-column">
                 <h4>Exports &amp; Admin</h4>
                 <ul>
-                    <li><a href="{{ '#' }}">Sellbrite Management</a></li>
-                    <li><a href="{{ '#' }}">Admin Dashboard</a></li>
-                    <li><a href="{{ '#' }}">Admin Security</a></li>
-                    <li><a href="{{ '#' }}">Description Editor (GDWS)</a></li>
+                    <li><a href="#">Sellbrite Management</a></li>
+                    <li><a href="{{ url_for('admin.dashboard') }}">Admin Dashboard</a></li>
+                    <li><a href="{{ url_for('admin.security_page') }}">Admin Security</a></li>
+                    <li><a href="#">Description Editor (GDWS)</a></li>
                 </ul>
                 {# Removed broken menu link: Admin Users (no such route exists) #}
             </div>
@@ -164,7 +164,7 @@
         </div>
     </footer>
 
-    <script src="{{ url_for('static', filename='js/main-overlay-test.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/main.js') }}"></script>
     <script src="{{ url_for('static', filename='js/analysis-modal.js') }}"></script>
 </body>
 </html>
