+++ templates/partials/footer.html
@@ -1,9 +1,9 @@
 <footer class="site-footer">
   <div class="footer-inner container">
     <nav class="footer-nav">
-      <a href="{{ url_for('artwork.home') }}">Home</a>
-      <a href="{{ url_for('artwork.artworks') }}">Artwork</a>
-      <a href="{{ url_for('exports.sellbrite_management') }}">Exports</a>
+      <a href="{{ url_for('home.home') }}">Home</a>
+      <a href="{{ url_for('home.artworks') }}">Artwork</a>
+      <a href="#">Exports</a>
       {% if session.get('username') == 'robbie' %}
         <a href="{{ url_for('admin.security_page') }}">Security</a>
       {% endif %}
