+++ templates/review_artwork.html
@@ -4,7 +4,7 @@
 <h1>Edit Listing: {{ slug }}</h1>
 <p class="page-description">Update artwork details and regenerate mockups before finalising.</p>
 <div class="main-content">
-  <img src="{{ url_for('artwork.finalised_image', filename=slug ~ '/' ~ slug ~ '.jpg') }}" alt="{{ slug }}" class="review-main-img" style="max-width:400px;display:block;margin-bottom:1rem;">
+  <img src="{{ url_for('static', filename='finalised-artwork/' ~ slug ~ '/' ~ slug ~ '.jpg') }}" alt="{{ slug }}" class="review-main-img" style="max-width:400px;display:block;margin-bottom:1rem;">
   <form method="post" action="{{ url_for('finalise.finalise', slug=slug) }}">
     <label for="title-input">Title</label>
     <input type="text" id="title-input" name="title" value="{{ title }}" required>
@@ -24,7 +24,7 @@
     </select>
     <div class="mockups">
       {% for i in range(1, 10) %}
-      <img src="{{ url_for('artwork.finalised_image', filename=slug ~ '/' ~ slug ~ '-MU-' ~ '%02d'|format(i) ~ '.jpg') }}" alt="Mockup {{ i }}">
+      <img src="{{ url_for('static', filename='finalised-artwork/' ~ slug ~ '/' ~ slug ~ '-MU-' ~ '%02d'|format(i) ~ '.jpg') }}" alt="Mockup {{ i }}">
       {% endfor %}
     </div>
     <button type="submit" name="action" value="finalise" class="btn btn-primary">Finalise Listing</button>
